Adds basic bullet functionality
Implements a basic bullet that can be fired by the player.

Adds a bullet sprite and scene, and makes the player able to shoot it during the idle_shoot animation.
The bullet moves in a straight line when instantiated, but its collision is not yet implemented.
Also deletes unused megaman sprite files.